modules/location_service/requirements.txt
kafka-python==2.0.2
marshmallow==3.7.1
psycopg2-binary==2.8.5